advanced search operations are added

+++ RESTService/src/main/java/boun/group9/webservice/helper/CommentChecker.java
@@ -51,9 +51,9 @@ public class CommentChecker {
 		return result;
 	}
 
-	public static String deleteComment(int commentID){
+	public static String deleteComment(int commentID, int userID){
 
-		String query = "DELETE FROM comments WHERE comments.id = " + commentID + ";" ;
+		String query = "DELETE FROM comments WHERE comments.id = " + commentID + " and comments.commented_by = "+ userID +";" ;
 		return query;
 	}
 
@@ -68,5 +68,40 @@ public class CommentChecker {
 		String query = "Update comments SET down_votes = down_votes + 1 WHERE comments.id = "+ commentID + ";";
 		return query;
 	}
+
+	public static ArrayList<Comments> getCommentsByCategory(int concertID, int category){
+		ArrayList<Comments> result = new ArrayList<Comments>();
+		Comments comment;
+		Users user;
+		String query = "SELECT Comments.id AS Comments_id , Comments.concert_id AS Comments_concert_id , Comments.commented_by AS Comments_commented_by , Comments.up_votes AS Comments_up_votes, Comments.down_votes AS Comments_down_votes, Comments.comment AS Comments_comment, Users.name AS Users_name, Users.email AS Users_email, Users.photo_path AS Users_photo_path FROM Comments INNER JOIN Users ON Users.id = Comments.commented_by WHERE Comments.concert_id="+concertID+" AND Comments.category = " + category + ";";
+		ResultSet rs;
+		try {
+			rs = Database.connect(query, Application.MODE_GET);
+			while(rs.next()) {
+				comment = new Comments();
+				user = new Users();
+				user.setName(rs.getString("Users_name"));
+				user.setPhoto_path(rs.getString("Users_photo_path"));
+				user.setEmail(rs.getString("Users_email"));
+				comment.setCategory(category);
+				comment.setId(rs.getInt("Comments_id"));
+				comment.setCommented_by(rs.getInt("Comments_commented_by"));
+				comment.setConcert_id(rs.getInt("Comments_concert_id"));
+				comment.setComment(rs.getString("Comments_comment"));
+				comment.setUp_votes(rs.getInt("Comments_up_votes"));
+				comment.setDown_votes(rs.getInt("Comments_down_votes"));
+				comment.setCommented_user(user);
+				result.add(comment);
+			}
+		}catch(SQLException ex) {
+			ex.printStackTrace();
+			return null;
+		}catch(NotSavedException ex) {
+			ex.printStackTrace();
+			return null;
+		}
+		return result;
+	}
+
 
 }
